refactor: make collection for amenities and make some validation and …

+++ Backend/app/Http/Controllers/api/AmenityController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\AmenityRequest;
+use App\Http\Resources\AmenityResource;
 use App\Models\Amenity;
 use Illuminate\Http\Request;
 
@@ -11,33 +12,46 @@ class AmenityController extends Controller
 {
     public function index()
     {
-        return Amenity::all();
+        return AmenityResource::collection(Amenity::all());
     }
 
     public function store(AmenityRequest $request)
     {
         $amenity = Amenity::create($request->validated());
-        return response()->json($amenity, 201);
+        return response()->json(new AmenityResource($amenity), 201);
     }
 
     public function show($slug)
     {
-        $amenity = Amenity::findOrFail($slug);
-        return response()->json($amenity);
+        $amenity = Amenity::where('slug', $slug)->first();
+
+        if (!$amenity) {
+            return response()->json(['error' => 'Amenity not found'], 404);
+        }
+
+        return response()->json(new AmenityResource($amenity));
     }
+
 
     public function update(AmenityRequest $request, $slug)
     {
-        $amenity = Amenity::findOrFail($slug);
+        $amenity = Amenity::where('slug', $slug)->first();
+
+        if (!$amenity) {
+            return response()->json(['error' => 'Amenity not found'], 404);
+        }
         $amenity->update($request->validated());
-        return response()->json($amenity);
+        return response()->json(new AmenityResource($amenity));
     }
 
     public function destroy($slug)
     {
-        $amenity = Amenity::findOrFail($slug);
+        $amenity = Amenity::where('slug', $slug)->first();
+
+        if (!$amenity) {
+            return response()->json(['error' => 'Amenity not found'], 404);
+        }
         $amenity->delete();
-        return response()->json(null, 204);
+        return response()->json(["message" => "Amenity deleted successfully"]);
     }
 }
-
